\name{sp500}
\docType{data}
\alias{sp500}
\title{A (mostly complete) subset of the SP500 with 250 data points}
\description{ This data set provides all components of the SP500 that are
  complete within the date range [2008-03-07, 2009-03-04] with returns suitable
  for running against the portfolio optimization routines.
  Note that the date range may periodically be updated to reflect current dates.
}
\usage{sp500}
\format{ A 250x491 zoo object of asset returns }
\source{ Yahoo! Finance }
\keyword{datasets}
