\name{sp500.subset}
\docType{data}
\alias{sp500.subset}
\title{A subset of the SP500 with 200 data points}
\description{ This data set provides a subset of the SP500 with returns suitable
for running against the portfolio optimization routines. }
\usage{sp500.subset}
\format{ A 200x75 zoo object of asset returns }
\source{ Yahoo! Finance }
\keyword{datasets}
